+++ release_notes.md
@@ -1,10 +1,11 @@
 # Release Notes
 
 ## 4.0.9
-
+-   APP-4866 - [transfomr] Add Support for "Severity" and "Status" fields to Case
 -   APP-4851 - [Util] Updated regex patterns for Playbook and TC variable to be more specific
 -   APP-4852 - [API] Send query params in body for GET requests
 -   APP-4846 - [transform] Add support for Batch API update which enables top-level associations.
+-   APP-4868 - [PREVIEW] [transform] add support for producing v3 api format
 
 
 ## 4.0.8